backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.3
